Updated extension methods namespaces
