+++ Kraken.Net.UnitTests/KrakenClientTests.cs
@@ -12,6 +12,7 @@ using CryptoExchange.Net.Objects;
 using Kraken.Net.Objects.Internal;
 using Kraken.Net.Clients;
 using Kraken.Net.Clients.SpotApi;
+using Kraken.Net.ExtensionMethods;
 using CryptoExchange.Net.Objects.Sockets;
 
 namespace Kraken.Net.UnitTests
