+++ Kraken.Net/ExtensionMethods/CryptoClientExtensions.cs
@@ -1,10 +1,7 @@
 ﻿using Kraken.Net.Clients;
 using Kraken.Net.Interfaces.Clients;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace CryptoExchange.Net.Clients
+namespace CryptoExchange.Net.Interfaces
 {
     /// <summary>
     /// Extensions for the ICryptoRestClient and ICryptoSocketClient interfaces
